+++ framework/core/src/http/discriminateAuthMethod.ts
@@ -1,7 +1,7 @@
 import { AnySchemaValidator } from '@forklaunch/validator';
-import { createHmac } from 'crypto';
 import { JWK, JWTPayload, jwtVerify } from 'jose';
 import { ParsedQs } from 'qs';
+import { createHmacToken } from './createHmacToken';
 import { isBasicAuthMethod } from './guards/isBasicAuthMethod';
 import { isHmacMethod } from './guards/isHmacMethod';
 import { isJwtAuthMethod } from './guards/isJwtAuthMethod';
@@ -89,15 +89,23 @@ export async function discriminateAuthMethod<
       type: 'hmac';
       auth: {
         secretKeys: Record<string, string>;
-        verificationFunction: (
-          method: string,
-          path: string,
-          body: string,
-          timestamp: string,
-          nonce: string,
-          signature: string,
-          secretKey: string
-        ) => Promise<boolean | undefined>;
+        verificationFunction: ({
+          body,
+          path,
+          method,
+          timestamp,
+          nonce,
+          signature,
+          secretKey
+        }: {
+          method: string;
+          path: string;
+          body?: unknown;
+          timestamp: string;
+          nonce: string;
+          signature: string;
+          secretKey: string;
+        }) => Promise<boolean | undefined>;
       };
     }
 > {
@@ -154,20 +162,33 @@ export async function discriminateAuthMethod<
       type: 'hmac' as const,
       auth: {
         secretKeys: auth.hmac.secretKeys,
-        verificationFunction: async (
-          method: string,
-          path: string,
-          body: string,
-          timestamp: string,
-          nonce: string,
-          signature: string,
-          secretKey: string
-        ) => {
-          const hmac = createHmac('sha256', secretKey);
-
-          hmac.update(`${method}\n${path}\n${body}\n${timestamp}\n${nonce}`);
-          const digest = hmac.digest('base64');
-          return digest === signature;
+        verificationFunction: async ({
+          method,
+          path,
+          body,
+          timestamp,
+          nonce,
+          signature,
+          secretKey
+        }: {
+          method: string;
+          path: string;
+          body?: unknown;
+          timestamp: string;
+          nonce: string;
+          signature: string;
+          secretKey: string;
+        }) => {
+          return (
+            createHmacToken({
+              method,
+              path,
+              body,
+              timestamp,
+              nonce,
+              secretKey
+            }) === signature
+          );
         }
       }
     };