+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 0.6.4
+
+### Patch Changes
+
+- Use more appropriate auth apis for use with auth around the app
+
 ## 0.6.3
 
 ### Patch Changes