+++ framework/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @forklaunch/common
 
+## 0.6.7
+
+### Patch Changes
+
+- slight hmac token creation signature change
+
+## 0.6.6
+
+### Patch Changes
+
+- Update packages and expose hmac key creation function
+
 ## 0.6.5
 
 ### Patch Changes