+++ framework/infrastructure/redis/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @forklaunch/infrastructure-redis
 
+## 0.4.9
+
+### Patch Changes
+
+- slight hmac token creation signature change
+- Updated dependencies
+  - @forklaunch/common@0.6.7
+  - @forklaunch/core@0.14.9
+
+## 0.4.8
+
+### Patch Changes
+
+- Update packages and expose hmac key creation function
+- Updated dependencies
+  - @forklaunch/common@0.6.6
+  - @forklaunch/core@0.14.8
+
 ## 0.4.7
 
 ### Patch Changes