+++ framework/validator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @forklaunch/validator
 
+## 0.10.7
+
+### Patch Changes
+
+- slight hmac token creation signature change
+- Updated dependencies
+  - @forklaunch/common@0.6.7
+
+## 0.10.6
+
+### Patch Changes
+
+- Update packages and expose hmac key creation function
+- Updated dependencies
+  - @forklaunch/common@0.6.6
+
 ## 0.10.5
 
 ### Patch Changes